WIB-31 Rename | src/letter로 경로 변경
