--- server/src/letter.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { letterGateway } from './letter.gateway';
-
-@Module({
-  providers: [letterGateway],
-})
-export class letterModule {}
